+++ components/app-sidebar.tsx
@@ -81,7 +81,7 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
             >
               <a href="#">
                 <IconInnerShadowTop className="!size-5" />
-                <span className="text-base font-semibold">Litesubs</span>
+                <span className="text-base font-semibold">liteSubs</span>
               </a>
             </SidebarMenuButton>
           </SidebarMenuItem>